+++ app/src/main/java/com/rockon999/android/leanbacklauncher/logging/LeanbackProto$Recommendation.java
@@ -48,7 +48,7 @@ public final class LeanbackProto$Recommendation extends ExtendableMessageNano<Le
 
     public void writeTo(CodedOutputByteBufferNano output) throws IOException {
         if (this.position != null) {
-            output.writeUInt32(1, this.position.intValue());
+            output.writeUInt32(1, this.position);
         }
         if (this.packageName != null) {
             output.writeString(2, this.packageName);
@@ -57,19 +57,19 @@ public final class LeanbackProto$Recommendation extends ExtendableMessageNano<Le
             output.writeString(3, this.tagName);
         }
         if (this.bucketId != null) {
-            output.writeUInt32(4, this.bucketId.intValue());
+            output.writeUInt32(4, this.bucketId);
         }
         if (this.notificationId != null) {
-            output.writeUInt32(5, this.notificationId.intValue());
+            output.writeUInt32(5, this.notificationId);
         }
         if (this.unnormalizedPriority != null) {
-            output.writeInt32(6, this.unnormalizedPriority.intValue());
+            output.writeInt32(6, this.unnormalizedPriority);
         }
         if (this.normalizedPriority != null) {
-            output.writeFloat(7, this.normalizedPriority.floatValue());
+            output.writeFloat(7, this.normalizedPriority);
         }
         if (this.score != null) {
-            output.writeFloat(8, this.score.floatValue());
+            output.writeFloat(8, this.score);
         }
         super.writeTo(output);
     }
@@ -77,7 +77,7 @@ public final class LeanbackProto$Recommendation extends ExtendableMessageNano<Le
     protected int computeSerializedSize() {
         int size = super.computeSerializedSize();
         if (this.position != null) {
-            size += CodedOutputByteBufferNano.computeUInt32Size(1, this.position.intValue());
+            size += CodedOutputByteBufferNano.computeUInt32Size(1, this.position);
         }
         if (this.packageName != null) {
             size += CodedOutputByteBufferNano.computeStringSize(2, this.packageName);
@@ -86,20 +86,20 @@ public final class LeanbackProto$Recommendation extends ExtendableMessageNano<Le
             size += CodedOutputByteBufferNano.computeStringSize(3, this.tagName);
         }
         if (this.bucketId != null) {
-            size += CodedOutputByteBufferNano.computeUInt32Size(4, this.bucketId.intValue());
+            size += CodedOutputByteBufferNano.computeUInt32Size(4, this.bucketId);
         }
         if (this.notificationId != null) {
-            size += CodedOutputByteBufferNano.computeUInt32Size(5, this.notificationId.intValue());
+            size += CodedOutputByteBufferNano.computeUInt32Size(5, this.notificationId);
         }
         if (this.unnormalizedPriority != null) {
-            size += CodedOutputByteBufferNano.computeInt32Size(6, this.unnormalizedPriority.intValue());
+            size += CodedOutputByteBufferNano.computeInt32Size(6, this.unnormalizedPriority);
         }
         if (this.normalizedPriority != null) {
-            size += CodedOutputByteBufferNano.computeFloatSize(7, this.normalizedPriority.floatValue());
+            size += CodedOutputByteBufferNano.computeFloatSize(7, this.normalizedPriority);
         }
         if (this.score == null) {
             return size;
         }
-        return size + CodedOutputByteBufferNano.computeFloatSize(8, this.score.floatValue());
+        return size + CodedOutputByteBufferNano.computeFloatSize(8, this.score);
     }
 }
